gathered the metrics collect logic in metrics_collector
