+++ main.go
@@ -53,9 +53,7 @@ func main() {
 		},
 	}
 
-	prometheus.MustRegister(collector.NewClusterHealth(logger, httpClient, esURL))
-	prometheus.MustRegister(collector.NewNodes(logger, httpClient, esURL, *esAllNodes))
-	prometheus.MustRegister(collector.NewIndices(logger, httpClient, esURL, *esAllNodes))
+	prometheus.MustRegister(collector.NewMetricsCollector(logger, httpClient, esURL, *esAllNodes))
 
 	http.Handle(*metricsPath, prometheus.Handler())
 	http.HandleFunc("/", IndexHandler(*metricsPath))
